Rename feature to `rgh-feature-descriptions`
